requirements.txt
altair==4.2.0
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
decorator==5.1.1
dnspython==2.2.1
entrypoints==0.4
filelock==3.8.0
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.12.0
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.15.0
loguru==0.6.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.23.2
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pinecone-client==2.0.13
protobuf==3.20.1
pyarrow==9.0.0
pydeck==0.8.0b1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
pyspark==3.0.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
rich==12.5.1
scikit-learn==1.1.2
scipy==1.9.1
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smmap==5.0.0
streamlit==1.12.2
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
transformers==4.21.2
typing-extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
win32-setctime==1.1.0
zipp==3.8.1
